+++ app/Http/Controllers/FeuilleReformeController.php
@@ -7,6 +7,7 @@ use App\Models\Log;
 use App\Models\Material;
 use App\Models\MaterialHistory;
 use Barryvdh\DomPDF\Facade\Pdf;
+use Illuminate\Support\Facades\Gate;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -51,22 +52,13 @@ class FeuilleReformeController extends Controller
                 ->latest('moved_at')
                 ->first();
 
-            if ($lastHistory) {
-                // Mettre à jour le dernier enregistrement
-                $lastHistory->update([
-                    'from_service_id' => $reforme->cedant_id,
-                    'to_service_id' => $reforme->cessionnaire_id,
-                    'moved_at' => now()
-                ]);
-            } else {
-                // Si aucun enregistrement n'est trouvé, vous pouvez créer un nouvel enregistrement
-                MaterialHistory::create([
-                    'material_id' => $reforme->material_id,
-                    'from_service_id' => $reforme->cedant_id,
-                    'to_service_id' => $reforme->cessionnaire_id,
-                    'moved_at' => now()
-                ]);
-            }
+            MaterialHistory::create([
+                'material_id' => $reforme->material_id,
+                'from_service_id' => $lastHistory->to_service_id,
+                'to_service_id' => $reforme->cessionnaire_id,
+                'moved_at' => now()
+            ]);
+
             // Créer le log
             Log::create([
                 'action' => 'create',
@@ -116,7 +108,13 @@ class FeuilleReformeController extends Controller
 
     public function reformedelete($id)
     {
+
         $reforme = FeuilleReforme::find($id);
+
+        // Check authorization using Gate
+        if (Gate::denies('delete', $reforme)) {
+            abort(403, 'Action non autorisée.');
+        }
         if ($reforme->delete()) {
             // Créer le log
             Log::create([
